Print newline before outputing info from `in`
Our output was getting appended to the end of the previous stdout string
output from the docker pull of our image.  This made it into one long
run-on sentence, and annoying.

Now we are prepending a '\n' before the first statement to ensure our
output starts on a new line.