Capture only string-valued k8s metadata in resource version metadata
Previously, we took all of the k8s resource's `.metadata {}` elements
and mapped them to name/value pairs to be plopped into our `metadata []`
listing emitted by `in`.

```json
{
  "version": { ... },
  "metadata": [
    { "name": "...", "value": "..." },
    { "name": "...", "value": "..." },
    { "name": "...", "value": "..." }
  ]
}
```

However, k8s metadata is not limited to simple string values like the
Concourse metadata is.  So before this commit, we could end up emitting
something like this:

```json
{
  "version": { ... },
  "metadata": [
    {
      "name": "...",
      "value": {
        "not": "a simple string"
      }
    }
  ]
}
```

for which Concourse would barf up a Go error attempting to bind a
non-string value.

This commit fixes that potential problem, by selecting only non-empty
string values to include in the metadata.

Closes #3