Support 'sensitive' resources
Added a new `sensitive` config option to both the `source` and `params`.
If a resource is declared "sensitive", then it will not be emitted to
the logs or be visible in the Concourse web UI.

By default, resources are not sensitive.

When used on the source config,
```yaml
resources:
  - name: my-k8s-resources
    source:
      sensitive: true
```
it will serve as the default for all `get` steps.

However, each `get` can override this,
```yaml
jobs:
  - name: view-resources
    plan:
      - get: my-k8s-resources
        version: every
        trigger: true
        params:
          sensitive: false
```

Closes #1