Keep entire response payload for filtering purposes during `check`
Small internal refactor of how we manage and keep the response
payload we get back from `kubectl`.  Previously, we immediately
filtered this down to the `items.metadata` for each resource that
came back.  We then passed that subset of the payload to the
various filtering methods.

However, this limited the filtering we could do to only what is
in the `metadata`.

Now, we keep all of the `.items[]` content from the response, and
pass that to our filtering functions.  This gives them the ability
to filter on things other than just what is in the `metadata` of
the resource.

No other features or capabilities have changed or been added in this
commit, it simply paves the way for additional filtering capabilities
(such as #9).