Wrap all custom jq filters in implicit '()'
Wrapped all custom jq filter expressions provided by the user's
configuration in `(...)`.  Prior to this, if the user's expression
included a pipe `|` operator, it would break the filter join.
The workaround would force the user to wrap all of his
expressions in a `(...)` in his resource config, which is a little
verbose and repetitive (not to mention annoying).

Now, it's handled automatically.  This doesn't cause any problems if the
user also provides his own `(...)` wrapping, so this is a
backwards compatible "fix".

Also polished the README section discussing the jq filter.  It was a
bit confusing, and had a few spelling and grammar errors.