Fix broken link in README
Found an old markdown-style link in the asciidoc `README.adoc`, and
corrected it.  The link is located in the documentation for the `source`
config section that mentioned configuring the cluster `namespace` you
want to target, and (tried) to link further down into the document where
it discusses how you can override the `namespace` as a `step` param.

Furthermore, the link used an implicit anchor (based on the heading
title).  This is no obvious that it is being used as an anchor, and is
brittle to potential title changes.

Added an explicit anchor to the section, and corrected the link syntax.